Bug fixes and Improvements
- Fixed: Updated to latest version of VLCKit fixing scrubbing bugs on
device
- Fixed: Value for Undefined Key crash if the image failed to load on
product view controller and delegate was not removed.

+++ PopcornTime/UI/iOS/Views/ProgressBar.swift
@@ -36,23 +36,23 @@ class ProgressBar: UIVisualEffectView {
         }
     }
     
-    var progress: CGFloat {
+    var progress: Float {
         get {
-            return CGFloat(progressSlider.value)
+            return progressSlider.value
         } set (value) {
-            progressSlider.value = Float(value)
+            progressSlider.value = value
         }
     }
     
-    var bufferProgress: CGFloat {
+    var bufferProgress: Float {
         get {
-            return CGFloat(bufferProgressView.progress)
+            return bufferProgressView.progress
         } set (progress) {
-            bufferProgressView.progress = Float(progress)
+            bufferProgressView.progress = progress
         }
     }
     
-    var scrubbingProgress: CGFloat {
+    var scrubbingProgress: Float {
         get {
             return progress
         } set (value) {
